pyproject.toml
[project]
name = "qrcodegenerator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "pillow>=12.1.0",
    "pyinstaller>=6.18.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-docx>=1.2.0",
    "qrcode>=8.2",
]
